srcs/game/requirements.txt
# 적용방법: pip install -r requirements.txt
# 만약 패키지 변경된 사항 있을 경우(추가설치 등): pip freeze > requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bidict==0.22.1
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
flake8==7.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
mccabe==0.7.0
msgpack==1.0.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
pyOpenSSL==23.3.0
python-engineio==4.8.2
python-socketio==5.11.0
pytz==2023.3.post1
redis==5.0.1
service-identity==24.1.0
setuptools==69.0.3
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzlocal==5.2
wsproto==1.2.0
zope.interface==6.1
